Fix bug in remove_double_header
If there is a second header row but it is blank for some columns, then
those columns will be parsed as not necessarily character vectors.
In other words, the presence of the blank for any column does not
interfere with the parsing of that column.
As a consequence, not only do those columns not need to be reparsed
after the double header row is removed, but the readr::parse_guess will
only work with character vectors, and so that would raise an error
because this function originally ran the parse_guess on everything().
Now, it just runs it on where(is.character) columns, and all other
columns are untouched.